create metrics subroutine
