+++ VegProcessor/utils.py
@@ -1508,37 +1508,6 @@ def save_variables_as_cogs(
     print("done exporting all variables.")
 
 
-def mask_sedflux_by_polygons(
-    polygons_path: str,
-    sedflux_paths: List[str],
-    years: List[int],
-    output_path: str,
-) -> None:
-    """Mask per-WY SEDFLUX data by subobjective polygon groups.
-
-    STUB. Mirrors the logic in cpra-metrics/delineate-subobjective-zones.py:
-    rasterize polygons grouped by ``PolyName`` onto the SEDFLUX grid, apply
-    each mask to the SEDFLUX data for each water year, and stack into a
-    ``(time, zone, y, x)`` Dataset written to ``output_path``.
-
-    Parameters
-    ----------
-    polygons_path : str
-        Path to the subobjective polygon shapefile. Must have ``PolyName``,
-        ``Subobjecti``, and ``Objective`` attribute columns.
-    sedflux_paths : list[str]
-        Per-water-year SEDFLUX zarr paths, in the same order as ``years``.
-    years : list[int]
-        Simulation water years, used as the ``time`` coordinate.
-    output_path : str
-        Destination NetCDF path.
-    """
-    raise NotImplementedError(
-        "mask_sedflux_by_polygons is a stub — port logic from "
-        "cpra-metrics/delineate-subobjective-zones.py"
-    )
-
-
 def process_netcdf_folder(
     input_folder: str,
     output_base_dir: str = "./data/",
